removing warnings and commenting
removing some warnings
commented on some functions

+++ Cproject/Src/mainmenu.c
@@ -80,10 +80,10 @@ void Mainmenu_Update(void)
 		currentElapseTime == 0;
 	}
 	
-	gamename_xpos = CP_Math_LerpFloat(-2000, (windowHeight / 2) - 100, lerp);
-	player_xpos = CP_Math_LerpFloat(-2000, windowWidth + 250, lerp);
+	gamename_xpos = CP_Math_LerpFloat(-2000, (float)((windowHeight / 2) - 100), lerp);
+	player_xpos = CP_Math_LerpFloat(-2000, (float)(windowWidth + 250), lerp);
 	ass_xpos = (player_xpos - 180);
-	player_return_xpos = CP_Math_LerpFloat(windowWidth+2000, -250, lerp);
+	player_return_xpos = CP_Math_LerpFloat((float)(windowWidth + 2000), -250, lerp);
 	ass_return_xpos = (player_return_xpos + 180);
 
 
@@ -93,13 +93,13 @@ void Mainmenu_Update(void)
 	CP_Settings_Fill(red);
 	CP_Settings_RectMode(CP_POSITION_CENTER);
 	//draw the draw rect in the middle of the anchor point //draw play button
-	CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2)+50, 200, 80);
+	CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 50, 200, 80);
 	//draw the draw 2nd rect in the middle of the anchor point //draw tutorial button
-	CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2) + 150, 200, 80);
+	CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 150, 200, 80);
 	//draw the draw 3rd rect in the middle of the anchor point //draw credit button
-	CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2) + 250, 200, 80);
+	CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 250, 200, 80);
 	//draw the draw 4th rect in the middle of the anchor point //draw exit button
-	CP_Graphics_DrawRect((windowWidth / 2), (windowHeight / 2) + 350, 200, 80);
+	CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 350, 200, 80);
 
 
 
@@ -108,16 +108,16 @@ void Mainmenu_Update(void)
 	//set font size
 	CP_Settings_TextSize(50.0f);
 	//text in the first rect 
-	CP_Font_DrawText("Play", windowWidth / 2, (windowHeight / 2)+50);
+	CP_Font_DrawText("Play", (float)(windowWidth / 2), (float)(windowHeight / 2) + 50);
 	CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 	//text in the 2nd rect
-	CP_Font_DrawText("Tutorial", windowWidth / 2, (windowHeight / 2) + 150);
+	CP_Font_DrawText("Tutorial", (float)(windowWidth / 2), (float)(windowHeight / 2) + 150);
 	CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 	//text in the 3nd rect
-	CP_Font_DrawText("Credit", windowWidth / 2, (windowHeight / 2) + 250);
+	CP_Font_DrawText("Credit", (float)(windowWidth / 2), (float)(windowHeight / 2) + 250);
 	CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 	//text in the 4nd rect
-	CP_Font_DrawText("Exit", (windowWidth / 2), (windowHeight / 2) + 350);
+	CP_Font_DrawText("Exit", (float)(windowWidth / 2), (float)(windowHeight / 2) + 350);
 	CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 
 
@@ -139,7 +139,7 @@ void Mainmenu_Update(void)
 		CP_Settings_ImageMode(CP_POSITION_CENTER);
 
 		//CP_Image_Draw(Gamename, windowWidth / 2, (windowHeight / 2) - 100, (float)CP_Image_GetWidth(Gamename), (float)CP_Image_GetHeight(Gamename), 255);
-		CP_Image_Draw(Gamename, windowWidth / 2, gamename_xpos, (float)CP_Image_GetWidth(Gamename)*2, (float)CP_Image_GetHeight(Gamename)*2, 255);
+		CP_Image_Draw(Gamename, (float)(windowWidth / 2), gamename_xpos, (float)CP_Image_GetWidth(Gamename) * 2, (float)CP_Image_GetHeight(Gamename) * 2, 255);
 
 		CP_Image_DrawSubImage(playerSS,
 			 //RENDERED POS AND SIZE
@@ -243,62 +243,62 @@ void Mainmenu_Update(void)
 	modulo will cap the alpha value and reset to 0 after it reached the amount its divided by
 	*/
 		//play button light up
-		if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) +50, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+		if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) +50, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 		{
 			CP_Settings_Fill(blue);
 			CP_Settings_RectMode(CP_POSITION_CENTER);
 			//draw the draw rect in the middle of the anchor point
-			CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2) +50 , 220, 100);
+			CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 50, 220, 100);
 			CP_Settings_Fill(black);
 			//set font size
 			CP_Settings_TextSize(70.0f);
 			//text in the first rect
-			CP_Font_DrawText("Play", windowWidth / 2, (windowHeight / 2)+50);
+			CP_Font_DrawText("Play", (float)(windowWidth / 2), (float)(windowHeight / 2) + 50);
 			CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 		}
 
 		//tutorial button light up
-		if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) + 150, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+		if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) + 150, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 		{
 			CP_Settings_Fill(blue);
 			CP_Settings_RectMode(CP_POSITION_CENTER);
 			//draw the draw rect in the middle of the anchor point
-			CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2) + 150, 220, 100);
+			CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 150, 220, 100);
 			CP_Settings_Fill(black);
 			//set font size
 			CP_Settings_TextSize(70.0f);
 			//text in the first rect
-			CP_Font_DrawText("Tutorial", windowWidth / 2, (windowHeight / 2) + 150);
+			CP_Font_DrawText("Tutorial", (float)(windowWidth / 2), (float)(windowHeight / 2) + 150);
 			CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 		}
 
 		//credit button light up
-		if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) + 250, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+		if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) + 250, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 		{
 			CP_Settings_Fill(blue);
 			CP_Settings_RectMode(CP_POSITION_CENTER);
 			//draw the draw rect in the middle of the anchor point
-			CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2) + 250, 220, 100);
+			CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 250, 220, 100);
 			CP_Settings_Fill(black);
 			//set font size
 			CP_Settings_TextSize(70.0f);
 			//text in the first rect
-			CP_Font_DrawText("Credit", windowWidth / 2, (windowHeight / 2) + 250);
+			CP_Font_DrawText("Credit", (float)(windowWidth / 2), (float)(windowHeight / 2) + 250);
 			CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 		}
 
 		//exit button light up
-		if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) + 350, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+		if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) + 350, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 		{
 			CP_Settings_Fill(blue);
 			CP_Settings_RectMode(CP_POSITION_CENTER);
 			//draw the draw rect in the middle of the anchor point
-			CP_Graphics_DrawRect(windowWidth / 2, (windowHeight / 2) + 350, 220, 100);
+			CP_Graphics_DrawRect((float)(windowWidth / 2), (float)(windowHeight / 2) + 350, 220, 100);
 			CP_Settings_Fill(black);
 			//set font size
 			CP_Settings_TextSize(70.0f);
 			//text in the first rect
-			CP_Font_DrawText("Exit", windowWidth / 2, (windowHeight / 2) + 350);
+			CP_Font_DrawText("Exit", (float)(windowWidth / 2), (float)(windowHeight / 2) + 350);
 			CP_Settings_TextAlignment(CP_TEXT_ALIGN_H_CENTER, CP_TEXT_ALIGN_V_MIDDLE);
 		}
 
@@ -306,7 +306,7 @@ void Mainmenu_Update(void)
 		//play button
 		if (CP_Input_MouseTriggered(MOUSE_BUTTON_LEFT) == 1)
 		{
-			if (IsAreaClicked((windowWidth / 2), (windowHeight / 2)+50, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+			if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2)+50, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 			{
 				CP_Engine_SetNextGameState(Level_Init, Level_Update, Level_Exit);
 
@@ -315,7 +315,7 @@ void Mainmenu_Update(void)
 		//tutorial button
 		if (CP_Input_MouseTriggered(MOUSE_BUTTON_LEFT) == 1)
 		{
-			if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) + 150, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+			if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) + 150, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 			{
 				CP_Engine_SetNextGameState(Level_Init, Level_Update, Level_Exit);
 
@@ -324,7 +324,7 @@ void Mainmenu_Update(void)
 		//credit button
 		if (CP_Input_MouseTriggered(MOUSE_BUTTON_LEFT) == 1)
 		{
-			if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) + 250, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+			if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) + 250, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 			{
 				CP_Engine_SetNextGameState(Level_Init, Level_Update, Level_Exit);
 
@@ -333,7 +333,7 @@ void Mainmenu_Update(void)
 		//exit button
 		if (CP_Input_MouseTriggered(MOUSE_BUTTON_LEFT) == 1)
 		{
-			if (IsAreaClicked((windowWidth / 2), (windowHeight / 2) + 350, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
+			if (IsAreaClicked((float)(windowWidth / 2), (float)(windowHeight / 2) + 350, 200, 80, CP_Input_GetMouseX(), CP_Input_GetMouseY()) == 1)
 			{
 				CP_Engine_Terminate();
 			}
